Create feature: flowlog
This feature is for taking notes while in a state of flow.
Switching to your notes and entering in what you are doing is cumbersome.
Instead, bind this to a single key, then log what you are doing.
The zshrc will remind you to process the local log file and then
delete it.